// src/app/shared/generated/model/discoveredToolInfo.ts
/**
 * onecx-ai-provider  service
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DangerLevel } from './dangerLevel';
import { DiscoveredToolAnnotations } from './discoveredToolAnnotations';
import { AgentMcpToolRule } from './agentMcpToolRule';


export interface DiscoveredToolInfo { 
    name?: string;
    description?: string;
    annotations?: DiscoveredToolAnnotations;
    autoDangerLevel?: DangerLevel;
    existingRule?: AgentMcpToolRule;
    orphaned?: boolean;
}